Navbar!
Next step is to make the buttons for the main page show on the navbar when a student is selected.

It's a bit slow because we're cdn-ing to bootstrap + plugins, jquery, and making ajax requests, running php etc but we can make it faster by perhaps installing some of those files.